Convert linked address keys into db object. Move bot profile, stream and refunds into own files. Check iterator errors.
